backend/requirements.txt
psycopg[binary]
djangorestframework-simplejwt
djangorestframework